@@ -1,45 +1,42 @@
 import React, { PropTypes, Component } from 'react';
 import _ from 'lodash';
+import { InputText, InputCombo, Multiselect, Select } from 'auth0-extension-ui';
+import { Button, Modal } from 'react-bootstrap';
+import { connect } from 'react-redux';
+import { reduxForm, Field, formValueSelector } from 'redux-form';
+import UserCustomFormFields from './UserCustomFormFields';
 
-import createForm from '../../utils/createForm';
-import { InputText, InputCombo, MultiSelect } from '../../components/Dashboard';
-
-export default createForm('user', class extends Component {
+class AddUserForm extends Component {
   static propTypes = {
-    fields: PropTypes.object.isRequired,
     initialValues: PropTypes.object,
     connections: PropTypes.array.isRequired,
     memberships: PropTypes.array.isRequired,
     createMemberships: PropTypes.boolean,
-    getDictValue: React.PropTypes.func
+    getDictValue: React.PropTypes.func,
+    hasSelectedConnection: PropTypes.string,
+    hasMembership: PropTypes.array,
+    onClose: React.PropTypes.func.isRequired,
+    handleSubmit: React.PropTypes.func.isRequired,
+    submitting: React.PropTypes.boolean,
+    customFields: React.PropTypes.array,
+    customFieldGetter: React.PropTypes.func.isRequired
   }
 
-  static formFields = [
-    'email',
-    'username',
-    'password',
-    'repeatPassword',
-    'connection',
-    'memberships'
-  ];
-
-  renderUsername(connections, connectionField, usernameField) {
-    const selectedConnection = _.find(connections, (conn) => conn.name === connectionField.value);
+  renderUsername(connections, hasSelectedConnection) {
+    const selectedConnection = _.find(connections, (conn) => conn.name === hasSelectedConnection);
     const requireUsername = selectedConnection && selectedConnection.options ? selectedConnection.options.requires_username : false;
-    if (!requireUsername) {
+    if (!requireUsername && (!this.props.initialValues || !this.props.initialValues.username)) {
       return null;
     }
 
     return (
-      <div className="custom-field">
-        <InputText field={usernameField} fieldName="username" label="Username" ref="username" />
-      </div>
+      <Field label="Username" name="username" component={InputText} />
     );
   }
 
-  renderMemberships(membershipsField, memberships, createMemberships) {
+  renderMemberships(hasMembership, memberships, createMemberships) {
     const allMemberships = _(memberships || [])
-      .concat(membershipsField.value)
+      .concat(hasMembership)
       .uniq()
       .sort()
       .value();
@@ -48,53 +45,148 @@ export default createForm('user', class extends Component {
     }
 
     return (
-      <div className="custom-field">
-        <div className="form-group">
-          <label>{this.props.getDictValue('memberships', 'Memberships')}</label>
-          <MultiSelect
-            allowCreate={createMemberships}
-            options={allMemberships.map(m => ({ value: m, label: m }))}
-            multi
-            {...membershipsField}
+      <div>
+        <label className="control-label col-xs-3" htmlFor="memberships">
+          {this.props.getDictValue('memberships', 'Memberships')}
+        </label>
+        <div className="col-xs-9" style={{ padding: '0 0 15px 7px' }}>
+          <Field
+            name="memberships"
+            id="memberships"
+            component={Multiselect}
+            loadOptions={(input, callback) => {
+              callback(null, {
+                options: allMemberships.map(m => ({ value: m, label: m })),
+                complete: true
+              });
+            }}
           />
         </div>
       </div>
     );
   }
 
-  renderConnections(field, connections) {
+  renderConnections(connections, customFields) {
     if (!connections || connections.length <= 1) {
       return null;
     }
 
-    const options = connections.map(conn => ({ value: conn.name, text: conn.name }));
+    const connectionUserField = _.find(customFields, { property: 'connection' });
+    const displayConnectionUserField = connectionUserField && !!connectionUserField.create;
 
-    return (
-      <div className="custom-field">
-        <InputCombo options={options} field={field} fieldName="connection" label="Connection" onChange={this.onConnectionChange} />
-      </div>
-    );
+    if (displayConnectionUserField) {
+      const options = connections.map(conn => ({ value: conn.name, text: conn.name }));
+      return (
+        <Field
+          label="Connection"
+          name="connection"
+          id="connection"
+          component={InputCombo}
+          options={options}
+          onChange={this.onConnectionChange}
+        />
+      );
+    }
+
+    return null;
+  }
+
+  renderPassword(customFields) {
+    const passwordField = _.find(customFields, { property: 'password' });
+    const displayField = !passwordField || !!passwordField.create;
+
+    if (displayField) {
+      return (
+        <div>
+          <Field
+            label="Password"
+            name="password"
+            type="password"
+            component={InputText}
+          />
+          <Field
+            label="Repeat Password"
+            name="repeatPassword"
+            type="password"
+            component={InputText}
+          />
+        </div>
+      );
+    }
+    return null;
+  }
+
+  renderEmail(customFields) {
+    const emailField = _.find(customFields, { property: 'email' });
+    const displayField = emailField && !!emailField.create;
+
+    if (displayField) {
+      return (
+        <Field
+          label="Email"
+          name="email"
+          component={InputText}
+        />
+      );
+    }
+    return null;
   }
 
   render() {
     const connections = this.props.connections;
-    const { fields, memberships, createMemberships } = this.props;
+
+    const {
+      submitting,
+      memberships,
+      createMemberships,
+      hasSelectedConnection,
+      hasMembership,
+      customFields
+    } = this.props;
+
+    const connectionUserField = _.find(this.props.customFields, { property: 'connection' });
+    const displayConnectionUserField = connectionUserField && !!connectionUserField.display;
 
     return (
-      <form className="create-user form-horizontal" style={{ marginTop: '30px' }}>
-        {this.renderMemberships(fields.memberships, memberships, createMemberships)}
-        <div className="custom-field">
-          <InputText field={fields.email} fieldName="email" label="Email" ref="email" />
-        </div>
-        {this.renderUsername(connections, fields.connection, fields.username)}
-        <div className="custom-field">
-          <InputText field={fields.password} fieldName="password" label="Password" type="password" ref="password" />
-        </div>
-        <div className="custom-field repeat-password">
-          <InputText field={fields.repeatPassword} fieldName="repeat-password" label="Repeat Password" type="password" ref="repeatPassword" />
-        </div>
-        {this.renderConnections(fields.connection, connections)}
-      </form>
+      <div>
+        <Modal.Body>
+          {this.props.children}
+          <div className="form-horizontal">
+            {this.renderMemberships(hasMembership, memberships, createMemberships)}
+            {this.renderEmail(customFields)}
+            {this.renderUsername(connections, hasSelectedConnection)}
+            {this.renderPassword(customFields)}
+            {this.renderConnections(connections, customFields) }
+            <UserCustomFormFields customFieldGetter={this.props.customFieldGetter} customFields={customFields} />
+          </div>
+        </Modal.Body>
+        <Modal.Footer>
+          <Button bsSize="large" bsStyle="transparent" disabled={submitting} onClick={this.props.onClose}>
+            Cancel
+          </Button>
+          <Button bsSize="large" bsStyle="primary" disabled={submitting} onClick={this.props.handleSubmit}>
+            {this.props.method || 'Create'}
+          </Button>
+        </Modal.Footer>
+      </div>
     );
   }
+}
+
+const reduxFormDecorator = reduxForm({
+  form: 'user'
 });
+
+// Decorate with connect to read form values
+const selector = formValueSelector('user');
+const connectDecorator = connect(state => {
+  const hasSelectedConnection = selector(state, 'connection');
+  const hasMembership = selector(state, 'memberships');
+
+  return {
+    hasSelectedConnection,
+    hasMembership
+  };
+});
+
+export default connectDecorator(reduxFormDecorator(AddUserForm));
