@@ -1,8 +1,9 @@
 import React from 'react';
 import { findDOMNode } from 'react-dom';
+import _ from 'lodash';
 
-import { SearchBar, UsersTable, UserPaginator } from './';
-import { Error, LoadingPanel, TableTotals } from '../Dashboard';
+import { LuceneSearchBar, UsersTable } from './';
+import { Error, LoadingPanel, TableTotals, SearchBar } from 'auth0-extension-ui';
 
 export default class UserOverview extends React.Component {
   static propTypes = {
@@ -11,32 +12,92 @@ export default class UserOverview extends React.Component {
     onPageChange: React.PropTypes.func.isRequired,
     error: React.PropTypes.object,
     users: React.PropTypes.array.isRequired,
-    total: React.PropTypes.number.isRequired,
-    pages: React.PropTypes.number.isRequired,
-    nextPage: React.PropTypes.number.isRequired,
-    loading: React.PropTypes.bool.isRequired
+    loading: React.PropTypes.bool.isRequired,
+    userFields: React.PropTypes.array.isRequired,
+    onColumnSort: React.PropTypes.func.isRequired,
+    sortOrder: React.PropTypes.number.isRequired,
+    sortProperty: React.PropTypes.string.isRequired
+  }
+
+  constructor(props) {
+    super(props);
+
+    this.searchOptions = _(this.props.userFields)
+      .filter(field => _.isObject(field.search) && field.search.filter && field.search.filter === true)
+      .map((field, index) => {
+        return {
+          title: field.label,
+          value: field.property,
+          filterBy: field.property,
+          selected: index === 0
+        };
+      })
+      .value();
+
+    this.defaultFilter = _.find(this.searchOptions, { selected: true });
+    this.state = {
+      searchValue: '',
+      selectedFilter: this.defaultFilter
+    };
+
+    this.onKeyPress = this.onKeyPress.bind(this);
+    this.onReset = this.onReset.bind(this);
+    this.onHandleOptionChange = this.onHandleOptionChange.bind(this);
   }
 
   onKeyPress = (e) => {
     if (e.key === 'Enter') {
-      this.props.onSearch(findDOMNode(this.refs.search).value);
+      e.preventDefault();
+      const query = e.target.value;
+      this.props.onSearch(query, this.state.selectedFilter.filterBy);
     }
   }
 
+  onReset() {
+    this.props.onReset();
+    this.setState({
+      searchValue: ''
+    });
+  }
+
+  onHandleOptionChange(option) {
+    this.setState({
+      selectedFilter: option
+    });
+  }
+
   render() {
-    const { loading, error, users, total, pages, nextPage } = this.props;
+    const { loading, sortProperty, sortOrder } = this.props;
     return (
       <div>
+        <div className="row">
+          <div className="col-xs-12 wrapper">
+            <Error message={this.props.error}/>
+          </div>
+        </div>  
+        <div className="row">
+          <div className="col-xs-12">
+            { (this.searchOptions && this.searchOptions.length > 0) ? (
+              <SearchBar
+                onReset={this.props.onReset}
+                enabled={!loading}
+                handleKeyPress={this.onKeyPress}
+                handleReset={this.onReset}
+                handleOptionChange={this.onHandleOptionChange}
+                searchOptions={this.searchOptions}
+                searchValue={this.state.searchValue} />
+            ) : (
+              <LuceneSearchBar
+                onReset={this.props.onReset}
+                onSearch={this.props.onSearch}
+                enabled={!loading} />
+            )}
+          </div>
+        </div>
         <LoadingPanel show={loading}>
           <div className="row">
-            <div className="col-xs-12 wrapper">
-              <Error message={error} />
-            </div>
-          </div>
-          <SearchBar onReset={this.props.onReset} onSearch={this.props.onSearch} enabled={!loading} />
-          <div className="row">
             <div className="col-xs-12">
-              <UsersTable loading={loading} users={users} />
+              <UsersTable loading={loading} users={this.props.users} userFields={this.props.userFields} onColumnSort={this.props.onColumnSort} sortOrder={sortOrder} sortProperty={sortProperty} />
             </div>
           </div>
         </LoadingPanel>
