@@ -19,7 +19,8 @@ export default connectContainer(class extends Component {
     databaseConnections: getUserDatabaseConnections(state),
     log: state.log,
     logs: state.user.get('logs'),
-    devices: state.user.get('devices')
+    devices: state.user.get('devices'),
+    settings: state.settings.get('record').toJS().settings
   });
 
   static actionsToProps = {
@@ -36,15 +37,30 @@ export default connectContainer(class extends Component {
     params: PropTypes.object,
     clearLog: React.PropTypes.func.isRequired,
     fetchLog: React.PropTypes.func.isRequired,
-    fetchUser: React.PropTypes.func.isRequired
+    fetchUser: React.PropTypes.func.isRequired,
+    getDictValue: React.PropTypes.func.isRequired
   }
 
   componentWillMount() {
     this.props.fetchUser(this.props.params.id);
   }
 
+  renderProfile(suppressRawData, user) {
+    if (suppressRawData) return null;
+
+    return (
+      <Tab eventKey={4} title="Profile">
+        <UserProfile loading={user.get('loading')} user={user.get('record')} error={user.get('error')} />
+      </Tab>
+    );
+
+  }
+
   render() {
-    const { user, databaseConnections, log, logs, devices } = this.props;
+    const { user, databaseConnections, log, logs, devices, settings } = this.props;
+    const userFields = (settings && settings.userFields) || [];
+    const suppressRawData = settings && settings.suppressRawData === true;
+
     return (
       <div className="user">
         <TabsHeader role={this.props.accessLevel.role} />
@@ -54,8 +70,10 @@ export default connectContainer(class extends Component {
             <div className="pull-right">
               <UserActions
                 user={user}
+                userFields={userFields}
                 databaseConnections={databaseConnections}
                 deleteUser={this.props.requestDeleteUser}
+                changeFields={this.props.requestFieldsChange}
                 resetPassword={this.props.requestPasswordReset}
                 changePassword={this.props.requestPasswordChange}
                 removeMfa={this.props.requestRemoveMultiFactor}
@@ -70,7 +88,7 @@ export default connectContainer(class extends Component {
         </div>
         <div className="row">
           <div className="col-xs-12">
-            <UserHeader loading={user.get('loading')} user={user.get('record')} error={user.get('error')} />
+            <UserHeader loading={user.get('loading')} user={user.get('record')} error={user.get('error')} userFields={userFields} />
           </div>
         </div>
         <div className="row user-tabs">
@@ -79,7 +97,9 @@ export default connectContainer(class extends Component {
               <Tab eventKey={1} title="User Information">
                 <UserInfo
                   loading={user.get('loading')} user={user.get('record')}
-                  memberships={user.get('memberships') && user.get('memberships').toJSON()} error={user.get('error')}
+                  memberships={user.get('memberships') && user.get('memberships').toJSON()}
+                  userFields={userFields}
+                  error={user.get('error')}
                 />
               </Tab>
               <Tab eventKey={2} title="Devices">
@@ -99,13 +119,12 @@ export default connectContainer(class extends Component {
                   error={logs.get('error')}
                 />
               </Tab>
-              <Tab eventKey={4} title="Profile">
-                <UserProfile loading={user.get('loading')} user={user.get('record')} error={user.get('error')} />
-              </Tab>
+              { this.renderProfile(suppressRawData, user) }
             </Tabs>
           </div>
         </div>
         <dialogs.DeleteDialog />
+        <dialogs.FieldsChangeDialog getDictValue={this.props.getDictValue} userFields={userFields} />
         <dialogs.EmailChangeDialog />
         <dialogs.PasswordResetDialog />
         <dialogs.PasswordChangeDialog />
