@@ -4,19 +4,44 @@ export default class UserHeader extends Component {
   static propTypes = {
     error: PropTypes.string,
     loading: PropTypes.bool.isRequired,
-    user: PropTypes.object.isRequired
+    user: PropTypes.object.isRequired,
+    userFields: React.PropTypes.array.isRequired
   }
 
   shouldComponentUpdate(nextProps) {
     return nextProps.user !== this.props.user || nextProps.loading !== this.props.loading;
   }
 
-  getEmail(user) {
-    if (!user.email) {
+  getName(user, userFields) {
+
+    const nameField = _.find(userFields, field => field.property === 'name');
+
+    if (nameField && _.isFunction(nameField.display)) {
+      /* Custom Name Field function, use that instead of email address */
+      return nameField.display(user);
+    }
+
+    return user.name || user.nickname || user.email;
+  }
+
+  getEmail(user, userFields) {
+    // Check for user.email right away to make sure the user has been initialized
+    if (!user.email) return <div></div>;
+
+    let email = user.email;
+
+    const emailField = _.find(userFields, field => field.property === 'email');
+
+    if (emailField && _.isFunction(emailField.display)) {
+      /* Custom Name Field function, use that instead of email address */
+      email = emailField.display(user);
+    }
+
+    if (!email || email.length === 0) {
       return <div></div>;
     }
 
-    return <span className="user-label user-head-email">{user.email}</span>;
+    return <span className="user-label user-head-email">{email}</span>;
   }
 
   render() {
@@ -25,6 +50,7 @@ export default class UserHeader extends Component {
     }
 
     const user = this.props.user.toJS();
+    const userFields = this.props.userFields || [];
 
     return (
       <div className="user-header">
@@ -39,9 +65,9 @@ export default class UserHeader extends Component {
             <div className="username-area">
               <h4>
                 <span className="name user-head-nickname">
-                  {user.name || user.nickname || user.email}
+                  { this.getName(user, userFields) }
                 </span>
-                {this.getEmail(user)}
+                { this.getEmail(user, userFields) }
               </h4>
             </div>
           </div>
