@@ -1,4 +1,5 @@
 export UserActions from './UserActions';
+export UserFieldsChangeForm from './UserFieldsChangeForm';
 export UserDevices from './UserDevices';
 export UserForm from './UserForm';
 export UserHeader from './UserHeader';
@@ -7,5 +8,5 @@ export UserOverview from './UserOverview';
 export UserProfile from './UserProfile';
 export UserInfo from './UserInfo';
 export UsersTable from './UsersTable';
-export SearchBar from './SearchBar';
+export LuceneSearchBar from './LuceneSearchBar';
 export UserPaginator from './UserPaginator';
