@@ -1,19 +1,21 @@
 import React, { Component, PropTypes } from 'react';
 import connectContainer from 'redux-static';
+import { Error } from 'auth0-extension-ui';
+import { Modal } from 'react-bootstrap';
 
-import { userActions } from '../../../actions';
+import { userActions, scriptActions } from '../../../actions';
 import { UserForm } from '../../../components/Users';
-import { Error, Confirm } from '../../../components/Dashboard';
 
 export default connectContainer(class extends Component {
   static stateToProps = (state) => ({
     userCreate: state.userCreate,
     accessLevel: state.accessLevel,
-    connections: state.connections
+    connections: state.connections,
   });
 
   static actionsToProps = {
-    ...userActions
+    ...userActions,
+    ...scriptActions
   }
 
   static propTypes = {
@@ -22,19 +24,16 @@ export default connectContainer(class extends Component {
     userCreate: PropTypes.object.isRequired,
     createUser: PropTypes.func.isRequired,
     getDictValue: PropTypes.func.isRequired,
-    cancelCreateUser: PropTypes.func.isRequired
+    cancelCreateUser: PropTypes.func.isRequired,
+    userFields: PropTypes.array.isRequired
   }
 
   shouldComponentUpdate(nextProps) {
-    return nextProps.userCreate !== this.props.userCreate || nextProps.connections !== this.props.connections || nextProps.accessLevel !== this.props.accessLevel;
+    return nextProps.userCreate !== this.props.userCreate || nextProps.connections !== this.props.connections || nextProps.accessLevel !== this.props.accessLevel || nextProps.userFields !== this.props.userFields;
   }
 
   onSubmit = (user) => {
     this.props.createUser(user);
-  }
-
-  onConfirm = () => {
-    this.refs.form.submit();
   }
 
   render() {
@@ -42,17 +41,27 @@ export default connectContainer(class extends Component {
     const connections = this.props.connections.toJS();
     const accessLevel = this.props.accessLevel.get('record').toJS();
 
+
+
     return (
-      <Confirm title="Create User" show={record !== null} loading={loading} onCancel={this.props.cancelCreateUser} onConfirm={this.onConfirm}>
-        <Error message={error} />
+      <Modal show={record !== null} className="modal-overflow-visible" onHide={this.props.cancelCreateUser}>
+        <Modal.Header closeButton={loading} className="has-border">
+          <Modal.Title>Create User</Modal.Title>
+        </Modal.Header>
+
         <UserForm
-          ref="form"
+          customFields={this.props.userFields || []}
+          customFieldGetter={field => field.create}
           connections={connections.records} initialValues={record}
           createMemberships={accessLevel.createMemberships}
-          memberships={accessLevel.memberships} onSubmit={this.onSubmit}
+          memberships={accessLevel.memberships}
           getDictValue={this.props.getDictValue}
-        />
-      </Confirm>
+          onClose={this.props.cancelCreateUser}
+          onSubmit={this.onSubmit}
+        >
+          <Error message={error} />
+        </UserForm>
+      </Modal>
     );
   }
 });
