@@ -1,12 +1,13 @@
 import React, { Component, PropTypes } from 'react';
 import connectContainer from 'redux-static';
+import { Error, Confirm } from 'auth0-extension-ui';
 
 import { userActions } from '../../../actions';
-import { Error, Confirm } from '../../../components/Dashboard';
 
 export default connectContainer(class extends Component {
   static stateToProps = (state) => ({
-    emailChange: state.emailChange
+    emailChange: state.emailChange,
+    settings: state.settings
   });
 
   static actionsToProps = {
@@ -27,9 +28,26 @@ export default connectContainer(class extends Component {
     this.props.changeEmail(this.refs.user.value, this.refs.email.value);
   }
 
+  renderConnection(connection, userFields) {
+    const connectionField = _.find(userFields, field => field.property === 'connection');
+
+    const displayConnection = !connectionField || (_.isBoolean(connectionField.edit) && connectionField.edit === true) || _.isObject(connectionField.edit);
+
+    return displayConnection ? <div className="form-group">
+      <label className="col-xs-2 control-label">Connection</label>
+      <div className="col-xs-9">
+        <input type="text" readOnly="readonly" className="form-control" value={connection} />
+      </div>
+    </div> : <div></div>;
+  }
+
   render() {
     const { cancelEmailChange } = this.props;
-    const { userId, connection, userEmail, userName, error, requesting, loading } = this.props.emailChange.toJS();
+    const { userId, customField, connection, userName, userEmail, error, requesting, loading } = this.props.emailChange.toJS();
+
+    const defaultEmailValue = customField ? customField.display(customField.user) : userEmail;
+
+    const userFields = _.get(this.props.settings.toJS(), 'record.settings.userFields', []);
 
     return (
       <Confirm
@@ -42,16 +60,11 @@ export default connectContainer(class extends Component {
         </p>
         <div className="row">
           <form className="form-horizontal col-xs-12" style={{ marginTop: '40px' }}>
-            <div className="form-group">
-              <label className="col-xs-2 control-label">Connection</label>
-              <div className="col-xs-9">
-                <input type="text" readOnly="readonly" className="form-control" value={connection} />
-              </div>
-            </div>
+            { this.renderConnection(connection, userFields) }
             <div className="form-group">
               <label className="col-xs-2 control-label">Email</label>
               <div className="col-xs-9">
-                <input ref="email" type="email" className="form-control" defaultValue={userEmail} />
+                <input ref="email" type="email" className="form-control" defaultValue={defaultEmailValue} />
               </div>
             </div>
             <input ref="user" type="hidden" readOnly="readonly" className="form-control" value={userId} />
