@@ -2,11 +2,12 @@ import React, { Component, PropTypes } from 'react';
 import connectContainer from 'redux-static';
 
 import { userActions } from '../../../actions';
-import { Error, Confirm } from '../../../components/Dashboard';
+import { Error, Confirm } from 'auth0-extension-ui';
 
 export default connectContainer(class extends Component {
   static stateToProps = (state) => ({
-    passwordChange: state.passwordChange
+    passwordChange: state.passwordChange,
+    settings: state.settings
   });
 
   static actionsToProps = {
@@ -27,9 +28,25 @@ export default connectContainer(class extends Component {
     this.props.changePassword(this.refs.password.value, this.refs.repeatPassword.value);
   }
 
+
+  renderConnection(connection, userFields) {
+    const connectionField = _.find(userFields, field => field.property === 'connection');
+
+    const displayConnection = !connectionField || (_.isBoolean(connectionField.edit) && connectionField.edit === true) || _.isObject(connectionField.edit);
+
+    return displayConnection ? <div className="form-group">
+      <label className="col-xs-2 control-label">Connection</label>
+      <div className="col-xs-9">
+        <input type="text" readOnly="readonly" className="form-control" value={connection} />
+      </div>
+    </div> : <div></div>;
+  }
+
   render() {
     const { cancelPasswordChange } = this.props;
     const { connection, userEmail, userName, error, requesting, loading } = this.props.passwordChange.toJS();
+
+    const userFields = _.get(this.props.settings.toJS(), 'record.settings.userFields', []);
 
     if (!requesting) {
       return null;
@@ -50,12 +67,7 @@ export default connectContainer(class extends Component {
                 <input type="text" readOnly="readonly" className="form-control" value={userEmail} />
               </div>
             </div>
-            <div className="form-group">
-              <label className="col-xs-2 control-label">Connection</label>
-              <div className="col-xs-9">
-                <input type="text" readOnly="readonly" className="form-control" value={connection} />
-              </div>
-            </div>
+            { this.renderConnection(connection, userFields) }
             <div className="form-group">
               <label className="col-xs-2 control-label">Password</label>
               <div className="col-xs-9">
