@@ -1,12 +1,12 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
+import { Pagination, TableTotals } from 'auth0-extension-ui';
 
 import { connectionActions, userActions } from '../../actions';
 
 import * as dialogs from './Dialogs';
 import TabsHeader from '../../components/TabsHeader';
 import { UserOverview, UserForm } from '../../components/Users';
-import { Pagination, TableTotals } from '../../components/Dashboard';
 
 import './Users.css';
 
@@ -26,7 +26,10 @@ class Users extends Component {
     getDictValue: PropTypes.func.isRequired,
     createUser: PropTypes.func.isRequired,
     fetchConnections: PropTypes.func.isRequired,
-    requestCreateUser: PropTypes.func.isRequired
+    requestCreateUser: PropTypes.func.isRequired,
+    settings: PropTypes.object.isRequired,
+    sortOrder: React.PropTypes.number.isRequired,
+    sortProperty: React.PropTypes.string.isRequired
   }
 
   constructor(props) {
@@ -45,8 +48,10 @@ class Users extends Component {
     this.props.fetchUsers('', false, page - 1);
   }
 
-  onSearch = (query) => {
-    this.props.fetchUsers(query);
+  onSearch = (query, filterBy) => {
+    if (query && query.length > 0) {
+      this.props.fetchUsers(query, false, 0, filterBy);
+    }
   }
 
   onReset = () => {
@@ -59,8 +64,29 @@ class Users extends Component {
     );
   }
 
+  onColumnSort = (sort) => {
+    this.props.fetchUsers('', false, 0, null, sort);
+  }
+
   render() {
-    const { loading, error, users, total, connections, userCreateError, userCreateLoading, accessLevel, nextPage, pages } = this.props;
+    const {
+      loading,
+      error,
+      users,
+      total,
+      connections,
+      userCreateError,
+      userCreateLoading,
+      accessLevel,
+      nextPage,
+      pages,
+      settings,
+      sortProperty,
+      sortOrder
+    } = this.props;
+
+    const userFields = (settings && settings.userFields) || [];
+
     return (
       <div className="users">
         <TabsHeader role={accessLevel.get('record').get('role')} />
@@ -75,7 +101,7 @@ class Users extends Component {
             : ''}
           </div>
         </div>
-        <dialogs.CreateDialog getDictValue={this.props.getDictValue} />
+        <dialogs.CreateDialog getDictValue={this.props.getDictValue} userFields={userFields} />
         <UserOverview
           onReset={this.onReset}
           onSearch={this.onSearch}
@@ -87,6 +113,10 @@ class Users extends Component {
           pages={pages}
           loading={loading}
           role={accessLevel.role}
+          userFields={userFields}
+          sortProperty={sortProperty}
+          sortOrder={sortOrder}
+          onColumnSort={this.onColumnSort}
         />
         <div className="row">
           <div className="col-xs-12">
@@ -95,11 +125,9 @@ class Users extends Component {
                 totalItems={total}
                 handlePageChange={this.onPageChange}
                 perPage={10}
-                currentPage={nextPage}
               /> :
               <TableTotals currentCount={users.length} totalCount={total} />
-            }
-            
+            }            
           </div>
         </div>
       </div>
@@ -119,7 +147,10 @@ function mapStateToProps(state) {
     connections: state.connections.get('records').toJS(),
     total: state.users.get('total'),
     nextPage: state.users.get('nextPage'),
-    pages: state.users.get('pages')
+    pages: state.users.get('pages'),
+    sortProperty: state.users.get('sortProperty'),
+    sortOrder: state.users.get('sortOrder'),
+    settings: state.settings.get('record').toJS().settings
   };
 }
 
