@@ -1,3 +1,2 @@
-export getUserAccessLevel from './getUserAccessLevel';
-export hasAccessLevel from './hasAccessLevel';
 export verifyUserAccess from './verifyUserAccess';
+export requireScope from './requireScope';
