@@ -1,3 +1,4 @@
+import _ from 'lodash';
 import nock from 'nock';
 import expect from 'expect';
 import Promise from 'bluebird';
@@ -8,7 +9,6 @@ import bodyParser from 'body-parser';
 import config from '../../../server/lib/config';
 import users from '../../../server/routes/users';
 import ScriptManager from '../../../server/lib/scriptmanager';
-import { getUserAccessLevel } from '../../../server/lib/middlewares';
 import { user, defaultUsers, defaultScripts, defaultConfig } from '../../utils/dummyData';
 
 describe('#users router', () => {
@@ -17,32 +17,55 @@ describe('#users router', () => {
   const fakeApiClient = (req, res, next) => {
     req.auth0 = {
       users: {
-        getAll: () => Promise.resolve({ users: defaultUsers }),
-        get: (options) => {
-          const id = parseInt(options.id, 10) - 1;
-          return Promise.resolve(defaultUsers[id]);
+        getAll: (options) => {
+          if (options.q && options.q.startsWith('(user_id:"1")')) {
+            return Promise.resolve({
+              users: _.filter(defaultUsers, user => user.user_id === 1)
+            });
+          }
+          if (options.sort) {
+            const sortParts = options.sort.split(':');
+            const order = sortParts[1] < 0 ? 'desc' : 'asc';
+            return Promise.resolve({
+              users: _.sortByOrder(defaultUsers, [sortParts[0]], [order])
+            });
+          }
+          return Promise.resolve({ users: defaultUsers })
         },
-        create: (data) => {
-          defaultUsers.push(data);
-          return Promise.resolve();
-        },
-        delete: () => {
-          defaultUsers.pop();
-          return Promise.resolve();
-        },
-        update: (options, data) => {
-          const id = parseInt(options.id, 10) - 1;
-          if (data.email) defaultUsers[id].email = data.email;
-          if (data.username) defaultUsers[id].username = data.username;
-          if (data.password) defaultUsers[id].password = data.password;
-          if (data.blocked !== undefined) defaultUsers[id].blocked = data.blocked;
-          return Promise.resolve();
-        },
-        deleteMultifactorProvider: () => Promise.resolve()
+        get:
+          (options) => {
+            const id = parseInt(options.id, 10) - 1;
+            return Promise.resolve(defaultUsers[id]);
+          },
+        create:
+          (data) => {
+            defaultUsers.push(data);
+            return Promise.resolve();
+          },
+        delete:
+          () => {
+            defaultUsers.pop();
+            return Promise.resolve();
+          },
+        update:
+          (options, data) => {
+            const id = parseInt(options.id, 10) - 1;
+            if (data.email) defaultUsers[id].email = data.email;
+            if (data.username) defaultUsers[id].username = data.username;
+            if (data.password) defaultUsers[id].password = data.password;
+            if (data.blocked !== undefined) defaultUsers[id].blocked = data.blocked;
+            if (data.app_metadata) _.assign(defaultUsers[id].app_metadata, data.app_metadata);
+            return Promise.resolve();
+          },
+        deleteMultifactorProvider:
+          (options) => options.provider !== 'badProvider' ?
+            Promise.resolve() :
+            Promise.reject(new Error('bad provider'))
       },
       deviceCredentials: {
         getAll: () => Promise.resolve([])
-      },
+      }
+      ,
       jobs: {
         verifyEmail: () => Promise.resolve([])
       }
@@ -56,47 +79,55 @@ describe('#users router', () => {
     next();
   };
 
-  const storage = {
-    read: () => Promise.resolve(storage.data),
-    data: {
-      scripts: {
-        access: defaultScripts.access,
-        filter: defaultScripts.filter,
-        create: defaultScripts.create,
-        memberships: defaultScripts.memberships,
-        settings: defaultScripts.settings
-      }
+  const defaultScriptData = {
+    scripts: {
+      access: defaultScripts.access,
+      filter: defaultScripts.filter,
+      create: defaultScripts.create,
+      memberships: defaultScripts.memberships,
+      settings: defaultScripts.settings
     }
   };
 
+  const storage = {
+    read: () => Promise.resolve(storage.data),
+  };
+
   const scriptManager = new ScriptManager(storage);
+  const oldGetCached = scriptManager.getCached;
+  const skipCache = name => scriptManager.get(name);
+  const settingsWithUserFields = ((ctx, callback) => {
+    var result = {
+      connections: ['conn-a', 'conn-b'],
+      dict: {
+        title: ctx.request.user.email + ' dashboard',
+        memberships: 'Groups'
+      },
+      css: 'http://localhost:3001/app/default.css',
+      userFields: [
+        {
+          property: "email",
+          label: "Email"
+        }
+      ]
+    };
+    callback(null, result);
+  }).toString();
 
   const app = express();
 
   app.use(bodyParser.json());
   app.use(bodyParser.urlencoded({ extended: false }));
-  app.use('/users', fakeApiClient, addUserToReq, getUserAccessLevel, users(storage, scriptManager));
+  app.use('/users', fakeApiClient, addUserToReq, users(storage, scriptManager));
+  const domain = new RegExp(config('AUTH0_DOMAIN'));
 
   before(() => {
-    const domain = new RegExp(config('AUTH0_DOMAIN'));
-
-    nock(domain)
-      .post('/dbconnections/change_password')
-      .reply(204);
-
-    nock(domain)
-      .get(/logs/)
-      .times(2)
-      .reply(200, []);
-
-    nock(domain)
-      .get(/logs/)
-      .times(2)
-      .reply(400, []);
-
     nock(domain)
       .post('/oauth/token')
       .reply(200, { ok: true, access_token: 'access_token' });
+
+    storage.data = _.cloneDeep(defaultScriptData);
+    scriptManager.getCached = oldGetCached;
   });
 
   describe('#List', () => {
@@ -116,15 +147,28 @@ describe('#users router', () => {
       const userFour = defaultUsers.pop();
 
       request(app)
-        .get('/users')
+        .get('/users?sortProperty=user_id&sortOrder=-1')
         .expect('Content-Type', /json/)
         .expect(200)
         .end((err, res) => {
           if (err) throw err;
-          expect(res.body).toEqual({ users: defaultUsers });
-
+          const targetUsers = _.sortByOrder(_.cloneDeep(defaultUsers), ['user_id'],['desc']);
           defaultUsers.push(userFour);
           defaultUsers.push(userFive);
+          expect(res.body).toEqual({ users: targetUsers });
+
+          done();
+        });
+    });
+
+    it('should attempt to filter users by user_id', (done) => {
+      request(app)
+        .get('/users?filterBy=user_id&search=1')
+        .expect(200)
+        .expect('Content-Type', /json/)
+        .end((err, res) => {
+          if (err) throw err;
+          expect(res.body).toEqual({ users: [defaultUsers[0]] });
           done();
         });
     });
@@ -138,8 +182,34 @@ describe('#users router', () => {
         .expect(200)
         .end((err, res) => {
           if (err) throw err;
-          expect(res.body.memberships).toEqual([ 'deptA' ]);
+          expect(res.body.memberships).toEqual(['deptA']);
           expect(res.body.user.user_id).toEqual(1);
+          done();
+        });
+    });
+
+    it('should return user`s record 2', (done) => {
+      request(app)
+        .get('/users/2')
+        .expect('Content-Type', /json/)
+        .expect(200)
+        .end((err, res) => {
+          if (err) throw err;
+          expect(res.body.memberships).toEqual(['deptA']);
+          expect(res.body.user.user_id).toEqual(2);
+          done();
+        });
+    });
+
+    it('should return user`s record 3', (done) => {
+      request(app)
+        .get('/users/3')
+        .expect('Content-Type', /json/)
+        .expect(200)
+        .end((err, res) => {
+          if (err) throw err;
+          expect(res.body.memberships).toEqual([]);
+          expect(res.body.user.user_id).toEqual(3);
           done();
         });
     });
@@ -171,7 +241,7 @@ describe('#users router', () => {
     it('should create new user', (done) => {
       const newUser = {
         email: 'user6@example.com',
-        memberships: [ 'deptA' ]
+        memberships: ['deptA']
       };
 
       request(app)
@@ -189,7 +259,7 @@ describe('#users router', () => {
     it('should return "The email address is required" error', (done) => {
       const newUser = {
         email: '',
-        memberships: [ 'deptA' ]
+        memberships: ['deptA']
       };
 
       request(app)
@@ -208,7 +278,7 @@ describe('#users router', () => {
         email: 'user7@example.com',
         password: 'password',
         repeatPassword: 'repeatPassword',
-        memberships: [ 'deptA' ]
+        memberships: ['deptA']
       };
 
       request(app)
@@ -225,7 +295,7 @@ describe('#users router', () => {
     it('should return "access denied" error', (done) => {
       const newUser = {
         email: 'user7@example.com',
-        memberships: [ 'deptB' ]
+        memberships: ['deptB']
       };
 
       request(app)
@@ -275,6 +345,10 @@ describe('#users router', () => {
 
   describe('#Change Password', () => {
     it('should change user`s password', (done) => {
+      nock(domain)
+        .post('/dbconnections/change_password')
+        .reply(204);
+
       request(app)
         .put('/users/1/change-password')
         .send({ password: 'password', confirmPassword: 'password' })
@@ -311,6 +385,7 @@ describe('#users router', () => {
 
   describe('#Change Username', () => {
     it('should change user`s name', (done) => {
+      storage.data = _.cloneDeep(defaultScriptData);
       request(app)
         .put('/users/1/change-username')
         .send({ username: 'name' })
@@ -335,7 +410,7 @@ describe('#users router', () => {
   });
 
   describe('#Change Email', () => {
-    it('should change user`s email', (done) => {
+    it('should change users email', (done) => {
       request(app)
         .put('/users/1/change-email')
         .send({ email: 'new-user1@example.com' })
@@ -498,6 +573,10 @@ describe('#users router', () => {
 
   describe('#Logs for user', () => {
     it('should return list of logs', (done) => {
+      nock(domain)
+        .get(/logs/)
+        .reply(200, []);
+
       request(app)
         .get('/users/1/logs')
         .expect(200)
@@ -508,7 +587,25 @@ describe('#users router', () => {
         });
     });
 
+    it('should return list of logs 2', (done) => {
+      nock(domain)
+        .get(/logs/)
+        .replyWithError('something bad happened');
+
+      request(app)
+        .get('/users/1/logs')
+        .expect(500)
+        .end((err, res) => {
+          expect(res.text).toMatch(/something bad happened/);
+          done();
+        });
+    });
+
     it('should return "bad request" error', (done) => {
+      nock(domain)
+        .get(/logs/)
+        .reply(400, []);
+
       request(app)
         .get('/users/1/logs')
         .expect(500)
@@ -525,6 +622,236 @@ describe('#users router', () => {
         .expect(400)
         .end((err) => {
           if (err) throw err;
+          done();
+        });
+    });
+  });
+
+  describe('#Remove Multifactor guardian', () => {
+    it('delete guardian', (done) => {
+      nock(domain)
+        .get(/1\/enrollments/)
+        .reply(200, [{ id: 1 }]);
+
+      nock(domain)
+        .delete(/enrollments\/1/)
+        .reply(204);
+
+      request(app)
+        .del('/users/1/multifactor/guardian')
+        .expect(204)
+        .end((err) => {
+          if (err) throw err;
+          done();
+        });
+    });
+
+    it('delete guardian attempt when enrollment already gone', (done) => {
+      nock(domain)
+        .get(/2\/enrollments/)
+        .reply(200, []);
+
+      request(app)
+        .del('/users/2/multifactor/guardian')
+        .expect(204)
+        .end((err) => {
+          if (err) throw err;
+          done();
+        });
+    });
+
+    it('bad request on guardian delete', (done) => {
+      nock(domain)
+        .get(/3\/enrollments/)
+        .reply(200, [{ id: 1 }]);
+
+      request(app)
+        .del('/users/3/multifactor/guardian')
+        .expect(404)
+        .end((err) => {
+          if (err) throw err;
+          done();
+        });
+    });
+
+    it('bad request on guardian get', (done) => {
+      request(app)
+        .del('/users/1/multifactor/guardian')
+        .expect(404)
+        .end((err) => {
+          if (err) throw err;
+          done();
+        });
+    });
+
+    it('should return "bad request" error bad provider', (done) => {
+      request(app)
+        .del('/users/1/multifactor/badProvider')
+        .expect(500)
+        .end((err) => {
+          if (err) throw err;
+          done();
+        });
+    });
+  });
+
+  describe('#User patch', () => {
+    it('patch user successfully', (done) => {
+      request(app)
+        .patch('/users/1')
+        .send({
+          app_metadata: {
+            someNewKey: 'someNewValue'
+          }
+        })
+        .expect(204)
+        .end((err) => {
+          if (err) throw err;
+          expect(defaultUsers[0].app_metadata.someNewKey).toEqual('someNewValue');
+          done();
+        });
+    });
+
+    it('should return "access denied" error', (done) => {
+      request(app)
+        .patch('/users/5')
+        .expect(400)
+        .end((err) => {
+          if (err) throw err;
+          done();
+        });
+    });
+
+    it('should return "not found" error', (done) => {
+      request(app)
+        .patch('/users/6')
+        .expect(404)
+        .end((err) => {
+          if (err) throw err;
+          done();
+        });
+    });
+
+    it('should return "access denied" error', (done) => {
+      request(app)
+        .patch('/users/2')
+        .expect(400)
+        .end((err) => {
+          if (err) throw err;
+          done();
+        });
+    });
+
+  });
+
+  describe('#userFields should call write hook', () => {
+    before(() => {
+      scriptManager.getCached = skipCache;
+      storage.data.scripts.settings = settingsWithUserFields;
+
+      storage.data.scripts.create = ((ctx, callback) => {
+        if (!ctx.userFields) throw Error('write hook should get userFields');
+
+        callback(null, { email: ctx.payload.email, password: ctx.payload.password, username: ctx.payload.username });
+      }).toString();
+    });
+
+    it('change-password', (done) => {
+      request(app)
+        .put('/users/1/change-password')
+        .send({ password: 'pwd1', confirmPassword: 'pwd1' })
+        .expect(204)
+        .end((err) => {
+          if (err) throw err;
+          expect(defaultUsers[0].password).toEqual('pwd1');
+          done();
+        });
+    });
+
+    it('change-email', (done) => {
+      request(app)
+        .put('/users/1/change-email')
+        .send({ email: 'new-user2@example.com' })
+        .expect(204)
+        .end((err) => {
+          if (err) throw err;
+          expect(defaultUsers[0].email).toEqual('new-user2@example.com');
+          done();
+        });
+    });
+
+    it('change-username', (done) => {
+      request(app)
+        .put('/users/1/change-username')
+        .send({ username: 'name2' })
+        .expect(204)
+        .end((err) => {
+          if (err) throw err;
+          expect(defaultUsers[0].username).toEqual('name2');
+          done();
+        });
+    });
+
+  });
+
+  describe('#userFields defined validation errors', () => {
+    before(() => {
+      scriptManager.getCached = skipCache;
+      storage.data.scripts.settings = settingsWithUserFields;
+
+      storage.data.scripts.create = ((ctx, callback) => {
+        callback(null, {});
+      }).toString();
+    });
+
+    it('change-password', (done) => {
+      request(app)
+        .put('/users/1/change-password')
+        .send({ password: 'pwd12', confirmPassword: 'pwd12' })
+        .expect(500)
+        .end((err, res) => {
+          expect(res.error).toMatch(/ValidationError: The password is required/);
+          done();
+        });
+    });
+
+    it('change-email', (done) => {
+      request(app)
+        .put('/users/1/change-email')
+        .send({ email: 'new-user3@example.com' })
+        .expect(500)
+        .end((err, res) => {
+          expect(res.error).toMatch(/ValidationError: The email is required/);
+          done();
+        });
+    });
+
+    it('change-username', (done) => {
+      request(app)
+        .put('/users/1/change-username')
+        .send({ username: 'name3' })
+        .expect(500)
+        .end((err, res) => {
+          expect(res.error).toMatch(/ValidationError: The username is required/);
+          done();
+        });
+    });
+  });
+
+  describe('#extra errors', () => {
+    before(() => {
+      scriptManager.getCached = skipCache;
+      storage.data.scripts.memberships = ((ctx, callback) => {
+        callback(new Error('intentional error'));
+      }).toString();
+    });
+
+    it('get-by-id', (done) => {
+      request(app)
+        .get('/users/1')
+        .expect(400)
+        .end((err, res) => {
+          expect(res.error).toMatch(/Error: intentional error/);
           done();
         });
     });
